+++ HISTORY.md
@@ -1,6 +1,6 @@
 #
 * bring deck.js up to date (https://github.com/imakewebthings/deck.js/commit/069f63294abe8c2bfd0e3c9b34d26090802c4f46)
-* add --style-theme and --transition-theme options
+* add --style and --transition options
 
 # 0.4.2 2012-09-26
 * require rack >= 1.4.1